requirements.txt
pyyaml
pytest
selenium==4.11.2